requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-url-filter==0.3.15